pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pymife"
version = "0.0.4"
authors = [
  { name="mechfrog88", email="kelzzin2@gmail.com" },
]
dependencies = [
    'pycryptodome >= 3.18.0',
    'gmpy2 >= 2.1.5'
]
description = "Python Functional Encryption Library"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
